requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
six>=1.9.0 # MIT
pytz>=2013.6 # MIT
PyYAML>=3.10 # MIT
requests>=2.14.2 # Apache-2.0
ndg-httpsclient>=0.4.2;python_version<'3.0' # BSD
PrettyTable<0.8,>=0.7.1 # BSD
docutils>=0.11 # OSI-Approved Open Source, Public Domain
